+++ src/valueswap_backend/src/utils/types.rs
@@ -241,7 +241,7 @@ pub(crate) struct CreateCanisterArgumentExtended {
 
 pub struct SwapParams {
     pub token1_name : String,
-    pub token_amount : u64,
+    pub token_amount : f64,
     pub token2_name : String,
     pub ledger_canister_id1 : Principal,
     pub ledger_canister_id2 : Principal
